+++ ft_strlen.c
@@ -1,9 +1,9 @@
 #include "ft_printf.h"
 
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
